requirements.txt
# Backend
python
flask

# Data manipulation
numpy
pandas

# ML
pytorch
torchvision

# Visualization
matplotlib
seaborn